Removed Knowledge class and refactored project to use Knowledge2
