+++ tetrad-lib/src/main/java/edu/cmu/tetrad/data/NumberObjectDataSet.java
@@ -77,7 +77,7 @@ import java.util.*;
  *
  * @author Joseph Ramsey
  * @see edu.cmu.tetrad.data.Variable
- * @see edu.cmu.tetrad.data.Knowledge
+ * @see edu.cmu.tetrad.data.Knowledge2
  */
 public final class NumberObjectDataSet
         implements DataSet, TetradSerializable {
