+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/PcStable.java
@@ -22,7 +22,8 @@
 package edu.cmu.tetrad.search;
 
 import edu.cmu.tetrad.data.IKnowledge;
-import edu.cmu.tetrad.data.Knowledge;
+import edu.cmu.tetrad.data.Knowledge2;
+
 import edu.cmu.tetrad.graph.EdgeListGraph;
 import edu.cmu.tetrad.graph.Graph;
 import edu.cmu.tetrad.graph.Node;
@@ -50,7 +51,7 @@ public class PcStable implements GraphSearch {
     /**
      * Forbidden and required edges for the search.
      */
-    private IKnowledge knowledge = new Knowledge();
+    private IKnowledge knowledge = new Knowledge2();
 
     /**
      * Sepset information accumulated in the search.
